+++ how-to/workspace-platform-starter/CHANGELOG.md
@@ -61,6 +61,7 @@
 - Added Dock monitors the `conditions-changed` and refreshes its contents if the condition was used
 - Added url property to splashScreenProvider so you can provide your own custom location for the html content
 - Change splash screen progress updates are sent using channels so they work cross domain
+- Added support for Snap SDK, add `snapProvider` to `customSettings`
 
 ## v14